add swift 4.2 support
