+++ settings.gradle
@@ -1,6 +1,7 @@
 include 'tests'
 
 include 'swift4.1'
+include 'swift4.2'
 
 rootProject.name = 'runtime-swift-ibm'
