+++ swift4.2/CHANGELOG.md
@@ -1,0 +1,12 @@
+# IBM Functions Swift 4.2 Runtime
+
+## 1.0.0
+Initial Swift 4.2 image
+  - Image name: `ibmfunctions/action-swift-v4.2`
+  - Kind is: `swift:4.2`
+
+Swift runtime version: [swift-4.2.1-RELEASE](https://swift.org/builds/swift-4.2.1-release/ubuntu16.04/swift-4.2.1-RELEASE/swift-4.2.1-RELEASE-ubuntu16.04.tar.gz)
+
+Packages included:
+  - [Watson SDK 1.1.0](https://github.com/watson-developer-cloud/swift-sdk/releases/tag/1.1.0)
+
