+++ swift4.2/Dockerfile
@@ -1,0 +1,6 @@
+# Dockerfile extends Apache OpenWhisk Swift image https://github.com/apache/incubator-openwhisk-runtime-swift/blob/master/core/swift42Action/Dockerfile
+FROM csantanapr/action-swift-v4.2
+
+# Add Pre-Installed Pacakges for IBM
+COPY _Whisk.swift /swiftAction/
+COPY Package.swift /swiftAction/
