+++ swift4.2/build.gradle
@@ -1,0 +1,14 @@
+ext.dockerImageName = 'action-swift-v4.2'
+apply from: '../gradle/docker.gradle'
+
+distDocker.dependsOn 'copyWhisk'
+distDocker.finalizedBy('cleanup')
+
+task copyWhisk(type: Copy) {
+    from '../swift4.1/spm-build/_Whisk.swift'
+    into '_Whisk.swift'
+}
+
+task cleanup(type: Delete) {
+    delete '_Whisk.swift'
+}
