+++ tests/src/test/scala/runtime/actionContainers/Swift41ActionContainerTests.scala
@@ -17,13 +17,156 @@
 
 package runtime.actionContainers
 
+import java.io.File
+
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
+import spray.json.{JsObject, JsString}
+import actionContainers.ResourceHelpers
 
 @RunWith(classOf[JUnitRunner])
 class Swift41ActionContainerTests extends SwiftActionContainerTests {
+
   override lazy val swiftContainerImageName = "action-swift-v4.1"
   override lazy val swiftBinaryName = "tests/dat/build/swift4.1/HelloSwift4.zip"
-  override lazy val partyCompileCodable = "tests/dat/build/swift4.1/SwiftyRequestCodable.zip"
-  override lazy val partyCompile = "tests/dat/build/swift4.1/SwiftyRequest.zip"
+  lazy val partyCompile = "tests/dat/build/swift4.1/SwiftyRequest.zip"
+  lazy val partyCompileCodable = "tests/dat/build/swift4.1/SwiftyRequestCodable.zip"
+
+  val httpCode = """
+       | import Dispatch
+       | func main(args:[String: Any]) -> [String:Any] {
+       |     var resp :[String:Any] = ["error":"getUrl failed"]
+       |     guard let urlStr = args["getUrl"] as? String else {
+       |         return ["error":"getUrl not found in action input"]
+       |     }
+       |     guard let url = URL(string: urlStr) else {
+       |         return ["error":"invalid url string \(urlStr)"]
+       |     }
+       |     let request = URLRequest(url: url)
+       |     let session = URLSession(configuration: .default)
+       |     let semaphore = DispatchSemaphore(value: 0)
+       |     let task = session.dataTask(with: request, completionHandler: {data, response, error -> Void in
+       |         print("done with http request")
+       |         if let error = error {
+       |             print("There was an error \(error)")
+       |         } else if let data = data,
+       |             let response = response as? HTTPURLResponse,
+       |             response.statusCode == 200 {
+       |             do {
+       |                 let respJson = try JSONSerialization.jsonObject(with: data)
+       |                 if respJson is [String:Any] {
+       |                     resp = respJson as! [String:Any]
+       |                 } else {
+       |                     resp = ["error":"Response from server is not a dictionary"]
+       |                 }
+       |             } catch {
+       |                 resp = ["error":"Error creating json from response: \(error)"]
+       |             }
+       |         }
+       |         semaphore.signal()
+       |     })
+       |     task.resume()
+       |     _ = semaphore.wait(timeout: .distantFuture)
+       |     return resp
+       | }
+     """.stripMargin
+
+  it should "support ability to use 3rd party packages like SwiftyRequest" in {
+    val zip = new File(partyCompile).toPath
+    val code = ResourceHelpers.readAsBase64(zip)
+
+    val (out, err) = withActionContainer() { c =>
+      val (initCode, initRes) = c.init(initPayload(code))
+      initCode should be(200)
+
+      val args = JsObject("message" -> (JsString("serverless")))
+      val (runCode, runRes) = c.run(runPayload(args))
+
+      runCode should be(200)
+      val json = runRes.get.fields.get("json")
+      json shouldBe Some(args)
+    }
+
+    checkStreams(out, err, {
+      case (o, e) =>
+        if (enforceEmptyOutputStream) o shouldBe empty
+        e shouldBe empty
+    })
+  }
+
+  it should "support ability to use escaping completion in Codable" in {
+    val zip = new File(partyCompileCodable).toPath
+    val code = ResourceHelpers.readAsBase64(zip)
+
+    val (out, err) = withActionContainer() { c =>
+      val (initCode, initRes) = c.init(initPayload(code, main = "mainCodable"))
+      initCode should be(200)
+
+      val (runCode, runRes) = c.run(runPayload(JsObject()))
+
+      runCode should be(200)
+      runRes.get.fields.get("greeting") shouldBe Some(JsString("success"))
+
+    }
+
+    checkStreams(out, err, {
+      case (o, e) =>
+        if (enforceEmptyOutputStream) o shouldBe empty
+        e shouldBe empty
+    })
+  }
+
+  // TODO
+  // skip for swift 4.2, it responds with 502 on init, stderr doesn't contain compile errors
+  // compile errors are contained in result
+  it should "log compilation errors" in {
+    val (out, err) = withActionContainer() { c =>
+      val code = """
+                   | 10 PRINT "Hello!"
+                   | 20 GOTO 10
+                 """.stripMargin
+
+      val (initCode, _) = c.init(initPayload(code))
+      initCode should not be (200)
+    }
+
+    checkStreams(out, err, {
+      case (o, e) =>
+        if (enforceEmptyOutputStream) o shouldBe empty
+        e.toLowerCase should include("error")
+    })
+  }
+
+  // TODO
+  // swift 4.2 exceptions executable exiting doesn't return error from web proxy or ends container
+  // the action times out
+  it should "return some error on action error" in {
+    val (out, err) = withActionContainer() { c =>
+      val code = """
+                   | // You need an indirection, or swiftc detects the div/0
+                   | // at compile-time. Smart.
+                   | func div(x: Int, y: Int) -> Int {
+                   |     return x/y
+                   | }
+                   | func main(args: [String: Any]) -> [String: Any] {
+                   |     return [ "divBy0": div(x:5, y:0) ]
+                   | }
+                 """.stripMargin
+
+      val (initCode, _) = c.init(initPayload(code))
+      initCode should be(200)
+
+      val (runCode, runRes) = c.run(runPayload(JsObject()))
+      runCode should be(502)
+
+      runRes shouldBe defined
+      runRes.get.fields.get("error") shouldBe defined
+    }
+
+    checkStreams(out, err, {
+      case (o, e) =>
+        if (enforceEmptyOutputStream) o shouldBe empty
+        if (enforceEmptyOutputStream) e shouldBe empty
+    })
+  }
 }
