+++ tools/travis/build.sh
@@ -24,8 +24,14 @@ docker tag openwhisk/nodejs6action nodejs6action
 
 TERM=dumb ./gradlew install
 
+# install new version docker
+curl -fsSL https://get.docker.com -o get-docker.sh
+sudo sh get-docker.sh
+docker version
+
 # Build runtime
 cd $ROOTDIR
 TERM=dumb ./gradlew \
 :swift4.1:distDocker \
+:swift4.2:distDocker \
 -PdockerImagePrefix=${IMAGE_PREFIX}
